Return queried hosts in timeout no response error.
Change queryExecutor do() function to wrap ErrTimeoutNoResponse and
ErrConnectionClosed so that the list of the hosts that have been
queried can be included in the error message. This is being done
to help debug an issue in which many timeout failures happen when
one node in the cluster is unavailable.